WSTEAM1-455: Render paragraph for Single Key Point.
